Stream decryption to receiver's filesystem
Before this change, receivers had to wait for the entirety of a download
file to be pulled into (browser) memory before getting sent through a
(decryption) stream that terminates at a point on their device's
filesystem. With this change, decrypting and saving to disk begins
immediately the server responds with the first chunks of their file.
Whenever possible.